requirements.txt
requests
requests_kerberos
load_dotenv
openai
langchain
langchain_openai
langchain_experimental
streamlit
streamlit>=1.28
langchain>=0.0.217
#openai>=1.2
duckduckgo-search
anthropic>=0.3.0
trubrics>=1.4.3
streamlit-feedback
langchain-community
chardet
faiss-cpu
scikit-learn
openpyxl
fitz
PyMuPDF
python-dotenv